Use flag info file from build directly
Context: now build system will create flag info file directly, there is no need for aconfigd to create flag info file at run time, it should just copy the RO flag info file in container to /metadata. In this change, we update aconfigd to do that.

In addition, split aconfigd tests into two parts. Proton collider specific test to its own test target. This is so that each test can be run within 20ms which is need to for them to be qualified as presubmit tests.

Change-Id: I4a938ce13be61e73d397fa62affaac5b265745b1